Retry installation if compilervars.sh is missing.
There seem to be some sporadic failures where the Intel installer
script returns successfully, but nothing is actually installed.  Asking
Travis to retry the build corrects the issue, so hopefully simply
retrying /inside/ the script will work, too.  The download files are
cached, so it should be much faster for future iterations.